requirements.txt
torch
torchvision
transformers
accelerate
pandas
numpy
tqdm
pillow
openpyxl
python-dotenv
openai
dashscope
streamlit
googlemaps
pyproj
requests
google-generativeai
clip @ git+https://github.com/openai/CLIP.git